requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
psycopg2-binary 
pydantic
python-dotenv
alembic
python-multipart
apscheduler
pydantic[email]